Add timedelta option, fix timezone issues.
Naive datetime objects are now universally interpreted as localtime
on the python client side. We had to do this because we need to
figure out UTC time for uuids on the python side when generating
from a datetime. Its awkward to have some things interpreted as
local time on the python side while others as localtime on the db
side. So all interpretation is done in python now.

We recommend using aware datetime objects in any case, such objects
have no ambiguity.